Minor re-org to improve documentation readability
